refactor: move mcp server file
